% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{construct_fm_index}
\alias{construct_fm_index}
\title{Construct FM Index}
\usage{
construct_fm_index(strings)
}
\arguments{
\item{strings}{Vector of strings to construct FM index from}
}
\description{
Construct FM Index
}
