% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fm_index_construct}
\alias{fm_index_construct}
\title{Construct FM Index}
\usage{
fm_index_construct(strings, case_sensitive = FALSE)
}
\arguments{
\item{strings}{Vector of strings to construct FM index from}

\item{case_sensitive}{Build case-sensitive index if TRUE}
}
\description{
Construct FM Index
}
