% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fm_index_find}
\alias{fm_index_find}
\title{Find query in FM Index}
\usage{
fm_index_find(query, index)
}
\arguments{
\item{query}{Strings to find in FM index}

\item{index}{Pointer to index created with \code{\link[=fm_index_construct]{fm_index_construct()}}}
}
\description{
Find query in FM Index
}
