% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fm_index_find}
\alias{fm_index_find}
\title{Find given queries in FM Index}
\usage{
fm_index_find(queries, index)
}
\arguments{
\item{queries}{Vector of strings to find in FM index}

\item{index}{Index created with \code{\link[=fm_index_construct]{fm_index_construct()}}}
}
\value{
A data frame with three columns. \code{pattern_index} is the index
of the query pattern, \code{library_index} is the index of the matching
string in the index, and \code{position} is the starting position of the
match in the indexed string. All indices are 1-based.
}
\description{
Find given queries in FM Index
}
\examples{
data("state")
index <- fm_index_construct(state.name, case_sensitive = FALSE)
# Find all states with "new" in their names
hits <- fm_index_find("new", index)
hits
# Show matching strings in library
state.name[hits$library_index]

hits <- fm_index_find("ar", index)
hits
state.name[hits$library_index]

}
\seealso{
Other FM Index functions: 
\code{\link{fm_index_construct}()},
\code{\link{fm_index_save}()}
}
\concept{FM Index functions}
