% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fm_index_save}
\alias{fm_index_save}
\title{Save FM Index}
\usage{
fm_index_save(index, path)
}
\arguments{
\item{index}{Pointer to index created with \code{\link[=construct_fm_index]{construct_fm_index()}}}

\item{path}{Path to save index to}
}
\description{
Save FM Index
}
